Convert the tomb routine to use a map marker (rob, me).
This fixes Mantis 1675: the map marker will expire as soon as it detects
that there is no monster within the tomb.  Also, the map marker uses
both a source and a target.  The former is needed to detect if it is a
monster-cast Tomb of Doroklohe or a tomb created by Zin's Imprison
ability; and the latter is used to detect the imprisoned monster if it
breaks out of the tomb early, so that any effects of entombment can
still be applied to it.  Entombment can also expire while the player is
off-level.

Accordingly, the "entombed" enchantment is removed.