Abyss code cleanup.
Clean up the code involved in generating and shifting the abyss. This is spadework in preparation for abyss vaults.

Fixes abyss shifts copying the area around the player to the centre of the abyss and not clearing the old area (irc conversation between Linley and dpeg confirms that the old behaviour of not clearing the duplicate shifted terrain is a bug).

Fixes abyssal runes being more likely to appear at the NW of the abyss level, particularly as time spent in the abyss increases.