Make map_cell store feat+cloud+item+mons instead of only the topmost one (BREAKS SAVES) (v3)
Changes in v3 (rob):
- update to current master
- fix colouring of out-of-LOS clouds
- fix reload crash due to env.cgrid

Changes in v2:
- Features and monsters should now be properly grayed

Currently map_cell only stores the topmost "thing" in a cell.
This is good enough for the console view, but not for more
sophisticated clients (e.g. tiles).

This commit refactors the code so that all player knowledge about a
cell (including detailed monster and item information) is represented
in map_cell, and so that map_cell is used to build the console view.

Save compatibility is broken since the new map_cell is serialized
instead of the old one.

This will allow, for instance, to load console savegames in the tiles
version, and still get proper tiles for the explored area.

Furthermore, this new map_cell will be the basis of the Crawl protocol
for the client/server split.

Signed-off-by: Robert Vollmert <rvollmert@gmx.net>