Trash direction_chooser special command repeat code.
Fixes issue #1681.

Entering junk at the "target self" y/n prompt was remaining
in the repeat buffer after casting thanks to the special
logic there. It would be possible to just flush the buffer
after targeting, but I think it's better if the repeat code
consistently does nothing but repeat keys. If we ever have
a good concept of "last command performed", this can be
rethought.

Command repeat replays keys and nothing