Fix hardcoded "classic_colours" option in tiles.
Instead of using item colour (which is set only at item create time and
depends on the colour options), use a random number stored with each
item.  Through the magic of member rearranging, this new value doesn't
take up extra space.  Also, rearrange the doll tiles so that reasonable,
non-unrandart and non-goofy tiles are contiguous so that we can use this
new random value to select one of them.

On the upside, more tiles are now used for dolls.  Also, this is a prereq
for a glyphs-in-tiles mode.