Rewrite DS mutation selection
Previously monstrous mutation selection was done in its own (poorly)
written function which made numerous presumptions on the structure
of the mutations in question (3 mutations per facet, etc.). Furthermore
it was becoming apparent that there were inadvertent selection bugs
which were resulting in conflicting mutations being selected that
would never actually appear (because the conflict forced the mut
to fail).

New code integrates the monstrous selection into the original
_select_ds_mutations in such a manner that is flexible enough
to allow for >3 mutation facets and a varying number of body slot
facets. Hopefully the new code also addresses the bugs arising
where DS are missing mutations ; testing seems to indicate this is
true.