Modularize get_tension and demonic guardian
Splits get_tension into get_monster_tension without affecting the
functionality of get_tension so that one can easily obtain the tension
value of an individual monster. This change facilitates the change
in guardian which allows for the number of guardians to be balanced
based on the current tension.

This is achieved by looping through and manually totalling the tension
values for each monster, at any given point if the tension is less
than random2(mutlevel * 5), remove the next monster we iterate to
that is friendly and a demonic guardian.

Also nerfs the spawn rate of demons quite significantly.