Per Mantis 1729, differentiate between types of monster glow.
There are now two kinds: glowing with light, and glowing with radiation.
Currently, all monsters with mutagenic corpses that glow are assumed to
be glowing with radiation ([very] ugly things, eyes of draining, and
glowing shapeshifters), while all other monsters that glow are assumed
to be glowing with light.

There are no game-visible changes from this, as the Corona spell
currently treats both types of glow equally.