Remove deprecated special-case code
Sapprovorous and Gourmand were special cased to be not allowed
as random mutations (as they should be). However, they already
have raritys of 0, which prevents them from being acquired
anyways. If mutate blocks of code are resulting in either
of these mutations, those blocks of code need adjusting, not
mutate()