Demonspawn conflicting mutations
Mantis item 1753: temporary mutations which conflict with a ds
mutation that is about to be gained will be removed by the ds
mutation, but that level of ds mutation is forever lost. This
can result (when a player has - for example - Frail 1 [-10%hp])
in being xl 27 but having only 2 of the 3 levels of a facet
(Robust 2 [+20%hp]) instead of all 3 (Robust 3 [+30%hp]).