Save Hell entry branch as well as level.
Placing entries to Hell outside of the main dungeon caused issues in
that it would store the depth of the level, but not the branch;
therefore, returning from Hell would drop you out in the main dungeon at
the equivalent branch.

Example: entering Hell from Lair:8 would cause you to exit to D:20, or
thereabouts.