A start at allowing doors to be locked.
The default requires that three runes be in the player's inventory
before the door can be opened. No messages are automatically provided,
but can be included.

Will be extended to allow for other items apart from runes; currently
has the capacity to call a function when the door is opened.