Don't identify mimics with autoexplore: #1199.
This makes the assumption that the current behaviour with unidentified
mimics is the correct one: if you create a mimic in WizMode (&Mmimic)
and then use Xt to teleport right next to it, and then move into it with
a "!a" inscribed weapon (or a ranged weapon), you won't be prompted.

Previous behaviour with autoexplore was to prompt before moving,
allowing you to cancel out, and allowing the mimic to be identified to
the player, but not actually in game, giving you a free turn and the
ability to move away.

Current behaviour is now the same as the example described in paragraph
one.