Add Nellie tile, elephant statue, fix jungle book. (Poor_Yurik)
New tile for Nellie and elephant statues by Poor_Yurik. I've also
tweaked the jungle_book vault so that not all of the elephant statues
are converted.

Finally, added the crumbled_column tile I made up a few months back.
It's not perfect, but I'm sure someone can find a use for it.