Use a "tgrid" so we don't have to look up traps or shops linearly.
find_trap() used to take over 10% CPU in sprint.

The main optimization here is bailing out early when grd() says there
is no trap -- just that cuts the effort by a factor of several hundred.

That makes making the loopup O(1) mostly moot, but I included it anyway
since tgrid costs just 11KB of memory.

To not risk breaking something for the release, if tgrid is invalid, the
list of traps/shops will be searched the old way.