Sprint tweaks: teleportation.
This commit should prevent all forms of teleportation. It doesn't
currently prevent teleportation control, but seeing as teleportation
shouldn't ever occur, this isn't an issue. Teleportitis mutation is
ignored, as are *Tele items, most specifically, rings of teleportation.

Seeing as teleportation is now useless, all such items (scroll of
teleportaton, ring of teleportation, ring of teleport control) are
marked as useless. I've left wand of teleportation as favoured, however,
as you should still be able to use it to get rid of undesirable foes.