Fix Aliza's macro crash (!lm Aliza type=crash 1)
The game triggered an assert when

1. The keyboard buffer had several queued keystrokes (not macro-expanded)

2. The first keystroke triggered a Lua macro that added expanded keystrokes to
   the *end* of the macro buffer -- adding expanded keys to the *front* of
   the macro buffer was always safe.

3. The macro buffer ended up looking like this: <keys-waiting-macroexpansion>
   <macroexpanded-keys>

(3) triggers the assert, since Crawl expects macroexpanded keys to always be at
the front of the macro buffer.

Fixed by using a secondary keybuffer to collect keys injected by Lua macros and
add them to the primary macro buffer only when it is empty.

Lua sendkeys and process_keys both add expanded keys to the end of the macro
buffer, which is probably not intended; they should presumably be adding keys
to the front of the macro buffer. This fix does not address that, however.

Also clear macro buffers when restarting after a game ends
(restart_after_game=yes)