Fix incorrect attribution for Awaken Forest.
The messages are not the best, but doing this right would require either
introducing a new kill method or some hacks, and that would be a bad idea
this close to the release.

The "awoken" version, suggested by due, looks a bit better but may puzzle
people due to a noun phrase split by "(damage: 8)\n ..."; if you think this
is bad please change this to my earlier "awakening".