[2007] Fix tmut spells not working with restart_after_game.
The player class doesn't initialise all its fields, and restart_after_game was
resetting it to a stack-initialised instance with garbage values for the
uninitialised fields, such as is_undead.

Fixed to use a static-initialised player object to reset 'you' on game reset.
Eventually, a better fix would be to explicitly initialise all player fields in
the constructor initialiser list.