Some improvements to can_go_straight.
Mainly due to the use of num_feats_between, it was often
returning false in situations where there obviously was
a straight path to the target.

This should mean that the somewhat broken monster pathfinding
is invoked less often, and hopefully monsters don't forget
about following the player so easily.