Activate markers early when loading level.
Fixes issue #2034.

Markers were being activated late for some messaging reasons, but
they need to be activated early so they can register listeners
that are interested in events generated during load. In particular,
before update_level() issues DET_TURN_ELAPSED for the time that has
passed while the player was off-level.

I propose that lua markers that now issue messages at inappropriate
times during activation issue them in some other way, for example
by registering a message for level_welcome_message() or a variant
of that when returning, or by registering a listener for
DET_ENTERED_LEVEL, which is issued close to the end of load().