Improve handling of monster foe memory.
Previously it was usually 0, and set to a positive value if the monster
was seeking, foe was out of LOS, and memory was at 0. This was probably
meant to catch the instant the foe left LOS, but could also fire in
different situations, e.g. when the player returned to a level by
a different staircase.

Foe memory is now set every time behaviour is evaluated for a seeking
monster with foe in sight, so it should be at a sensible value when
the foe goes out of LOS. Also, memory is properly reduced when the
player returns from a different level.

This should fix monsters heading straight for the player when they
return to a level, regardless of time elapsed. They'll still head
for the player if foe memory is positive and the player fails a
stealth check.

This does not fix monsters not doing a particularly good job of finding
the player: They don't use pathfinding, hence are likely to get stuck
in corners if the level is not without inside corners.