Add a "demonic tentacle" monster, refactor tentacle code to support it
Add a demonic tentacle monster, suitable for use with the malign
gateway spell. This is a tentacle with a fixed base position, and no
monster acting as the "base" in the way the main kraken body does for
kraken tentacles.

Refactor some of the tentacle code to handle broken assumptions (there
is no "head" monster in this case, so the tentacle end is what we pay
attention to for damage synch, attitude checks etc.).