Make shafts and hatches choose the safest adjacent destination in slime.
The destination is fudged to the adjacent cell with fewest slime walls next
to it. With the bubble layout, this means usually not directly next to a wall,
and rarely (never?) more than four.

This doesn't help when a slime level has corridors (by digging, Dissolving or
vault placement, e.g.).