Rename chars used as numbers to int8_t/uint8_t.  Fix some other type usage.
This should help against the signed char problems, and is good for
code readability.  Now, if you have a char, it's either an untyped
in-memory byte, or a symbol inside a string.  Small numbers are instead
[u]int8_t, ints, an enum type, or, in so many cases, bools.

I didn't touch any of the tiles code, as it's currently broken and I don't
want to risk making it unbroken harder.