Fix crash with off-level map viewing. (#2219)
The game would crash if the cursor position was outside map bounds when
using []. This can happen when the player only knows a small part of the
map, which is always centered to avoid leaking information.

Not 100% sure this is the cause for the crash in issue #2219, but it's
quite likely.