Fix vetoed levels messing up previous level's connectivity data.
_fixup_interlevel_connectivity was marking the previous level's
regions as connected based on the currently generated one and
not undoing these changes if it failed later.

On future tries, it was then operating on bad data, which could
well explain the Orcish Mines disconnectedness problems we've
been seeing.

Now only change the previous level's data once the checks are
passed successfully. The code still relies on the level not
being vetoed after the call to _fixup_interlevel_connectivity.