Add variants of existing Temple designs with fewer altars.
I hope I'm not stepping on anyone's toes here. I like those designs,
or I wouldn't have copied them.

As a result of all the recent Temple map changes, the chances for
getting one of the 12-altar maps is now much lower: 58% (was 81%)

Chance for <  8 altars: 11% (was: 4%)
Chance for < 10 altars: 23% (was: 5%)
Chance for < 12 altars: 32% (was: 6%)

Expected (average) number of altars: 10.7 (was 11.8)