Get rid of extra chance for special unrandart generation.
Fixes issue #2361.

Item generation now doesn't distinguish between special and
non-special artefacts. Special artefacts are generated
through the same path as non-special ones.

It would be possible to increase odds of artefact generation to
compensate, but I think with loot levels on the rise over the last
few releases, that's really not necessary.

This also unifies abyss and pandemonium handling: Unrands are
not generated randomly for either. The only exception is that
unrands that are not generated randomly but that were left in
the abyss may show up in the abyss again.