add ambient noise levels to branches
noise loudness is increased or decreased by a branch-dependent amount to
simulate having background noise. hive is very noisy and crypt is very
quiet, and a couple other branches are in between... we might want to
give values to a couple other ones, but couldn't really think of good
ideas for that off the top of my head.

i'm not sure how much this actually affects with our current noise
system, but i think it's a reasonable concept to build on.

this also gives a bonus/malus to stealth, the size of which is certainly
up for tweaking (with the current numbers, hive is +40 and crypt is
-40).