"--edit-save Joe repack" that defrags the save file.
It's mostly so you can see how much of the file was lost to due to waste.
If everything is working as designed, waste should never exceed the size of
compressed global chunks (chr, you, tc, kil, ...) plus at most two levels
-- that is, the biggest set of chunks rewritten between commits.

If you know a case when the waste explodes in size, tell me!