Remove any specific numbers from monster enums.
The MONS_UNUSED_XXX entries may be unsightly but are safer, and can be removed
if we decide to sort the enums later.  I didn't do so since that would break
save compat (or require a lot of shims).

The use of MONS_PROGRAM_BUG for 250-headed hydras was something the compiler
ought shout at people for (an incompatible type); removed.  I used 27 only
because I'm too lazy to add proper error catching there right now.