make statues block opacity_no_trans
this means that monsters behind statues won't make you feel unsafe
(mantis #832), and it also blocks apportation from working through
granite statues (and also probably some other things like that... is
this what we want? the new granite statues on zot:5 are meaningless
otherwise). people should test various things with this to make sure
the interactions are all still as we want them.