A spotty vault connector for the Mines. (#2349)
This is quite experimental; I'm not very familiar with the
dungeon generation code. Please test.

I think it's not too expensive, but I didn't really take care
of that. It would seem the selection of random seeds could
be improved.

Currently it just overrides the join_the_dots-connector in the
Orcish Mines (with fallback). It's not used for other spotty
levels.