Give experience at the end of monster_die().
This might switch message order a bit, but it didn't seem like a bad
change in my tests.

This should fix the various monster cleanup issues we have when the game
is interrupted at the stat gain prompt. E.g. corpse and monster inventory
weirdness.

One concrete bug that's fixed is that after blowing up an ogre with
disintegration at XL 1 and hanging up at the stat gain prompt, you'll
get an unlinked item error for the ogre's weapon.