Make equipped zombified deep elf blademasters able to dual-wield.
This applies to both equipped zombies and equipped intact enslaved
souls, and is accomplished by adding a flag to force the monster to be
treated as a dual-wielder, and setting it for the zombified type.

Note that this should also work with any other future monsters that are
dual-wielders, but which have non-dual-wielding base monster types.