Fix levitation removal prevention logic.
Part of issue #2396.

For one, it's a lot simpler and more correct now, and takes into account the
updated uncancellable levitation logic.

Then, it no longer duplicates the scrambling-free-logic. This is really not
something the interface should know about, and if this is done, the check
should not be duplicated. Finally, it now asks the player instead of
outright denying. Aside from scrambling free, there are situations where you
might want to stop levitation anyway (e.g. scramblable lava with enough rF+
if the item has some other bad properties).