Remove shatter damage to transformed casters.
This prevents synergy between the two highest Earth/Tmut spells
(lots of damage when shattering in statue form), and considering
the statue form slowness, it's not really necessary with respect
to balance.

Also, it's hard to explain why shatter should not hurt an
untransformed caster.

If shatter were too strong, the other option would be to damage
the player even when untransformed.