Don't treat contaminated chunks as clean for ghouls.
Ghouls have a greater chance of positive effects from contaminated
chunks, and the amulet of the gourmand makes chunks contaminated, in any
case, so this seems to be an oversight.

Also, the prefer_safe_chunks option no longer applies to undead, as
ghouls are the only undead that eat chunks, and both clean and
contaminated chunks are safe for them.