Fix passwall and teleportation, trampling. (#2582)
Passwall is now interrupted exactly like stair delays, and
trampling interrupts like teleportation and nets.

Stair delays will now be interrupted properly when trampled,
instead of failing because the player position changed.
You can no longer be trampled from one stair to the next
and ascend that instead.

It's not completely clear to me that this is all implemented
optimally, but if throwing nets pretend to be teleportation,
surely trampling can, too.