Partially implement spirits, though not fully.
They still need to summon a top-tier holy monster when they day, and
need better indications of how far along they are to actually carking
it; currently their message shifts from "fading away" to "mostly faded
away" when they're less than 20 or so turns from completely going.

They're implemented as constantly fleeing, but are quite slow. You can't
actually corner them ever, which may be a problem from an interface
point of view. It's not necessarily the best implementation, but it did
the trick... might have to change. :)