Fix a few instances of uniques' having the wrong species type.
Specifically, their species type was set to their unique monster type,
so that of a corpse drop or zombification was forced, you'd get e.g.
Antaeus the Antaeus corpse.

Changes, based on monster descriptions: Antaeus is a titan, Dissolution
is a jelly, Grinder is a shadow imp, and Roxanne is a statue.

Not fixed yet: the Laernaean hydra, the Serpent of Hell, the royal
jelly, Mnoleg, Lom Lobon, Cerebov, Gloorx Vloq, Geryon, Dispater,
Asmodeus, and Ereshkigal.