Ring of charms: disallow transforms, try another formula for success.
The former would allow charming spiders while in spider form, liches
while necromutated, and so on.

The latter in turn made the ring useless for any monster with
non-negligible HD.  Even an xl 27 merfolk had minimal changes with
base 9 HD merfolk.  This formula is perhaps way too generous, but
it includes a hard cap.  Improvements welcome.