Implement fear as a monster spell, player status.
The opposite of mesmerisation: instead of not being able to move away
from the creature in question, you are unable to move closer to them.
This is useful for ranged fighters, but useless with monsters that will
move closer to you. One merely needs to hold . until they arrive, or hit
them with spells, etc.

The range could be tweaked slightly. I've set it up as olddist >=
newdist, which means that you get the following effect (where 5 is a
monster causing fear, : are squares you cannot move into, . as normal):

:::::::..
:::::::..
:::5::@..
:::::::..
:::::::..

Switching to olddist > newdist results in:

::::::...
::::::...
:::5::@..
::::::...
::::::...

Square LOS may also make a difference in this case.

No monster currently uses it, but giving them the spell "cause fear"
works as imagined.