Make radioactive glow different from light glow in-game.
Specifically, allow (very) ugly things to have a chance of mutating when
adjacent to any monster that glows with radiation.  They will still have
a higher chance of mutating when adjacent to their own kin, however.

Note that the only monsters that glow with radiation other than (very)
ugly things are glowing shapeshifters in their original form, shining
eyes, and orbs of fire, so not much should currently be affected.