Simplify the previous stat preservation routines for zombies.
The use of the original monster type for the stats and the base type is
now the default, so the parameter is removed.  Additionally, the
original monster type is now properly passed in to the original
mgen_data() calls where applicable, so define_zombie() no longer needs
to effectively be called twice for spells and abilities that create
zombified types.  (Yred's Enslave Soul still needs to call it, since it
bypasses the monster placement code entirely when it recreates the
monster.)