Fix wand usage for monsters to match logic elsewhere.
According to comments in mon-data.h as well as usage, only monsters with
MONUSE_WEAPONS_ARMOUR or greater, which includes MONUSE_MAGIC_ITEMS, are
supposed to be able to use wands (and now rods). Simplify the logic in
_handle_wand() accordingly.  The only monsters that can be generated
with rods, deep dwarf artificers, have MONUSE_WEAPONS_ARMOUR, so they
won't be affected.