change the mephitic cloud message
mephitic is an explosion and makes as much noise as a fireball. (Mantis 2616)
change "The beam expands into a vile cloud!" into "The beam explodes..."
change "You hear a gentle \'poof\'." by "You hear a loud \'bang\'!"
Add an exclamation mark to all the explosion messages that were missing it.