Sif Muna's destroy spellbook conduct
Destroying a spellbook by forgetting a spell from it or tossing it in lava cost piety and penance.
The cost is based on the highest spell level from the book: maxlevel + 5
For books gifted by Sif or featuring his name, the penance is doubled.
The conduct description is only "destroy spellbooks". I don't know how to explain the gift and name higher cost without making a huge line