Ashenzari: recognize curses on sight.
Note that I modify a const object here -- which is a bad thing to do; just
remember how we chastised blue_anna for doing that.  Too bad, fixing that
would require not just merely removing the qualifier from several layers of
functions, but also rewriting the stack_iterator.  Fortunately, when we had
mirror damage, it inflicted damage to a const monster which could remove
(kill it), wake it up, spawn jellies via passive damage, etc -- but here,
it's just setting a flag that cannot invalidate pointers or so on.