Remove poison resistance from abominations again.
It was originally added so that their undead versions would be properly
poison-resistant, and so that the two would have identical resistances
in order to avoid confusion, but now that poison resistance is added for
all undead types, and demonic abominations can no longer be confused
with undead abominations in terms of color or damage messages ("wounded"
vs. "damaged"), it seems to no longer be necessary.  Besides, enough
demonic monsters resist poison as it is.