Clean up Tiamat's color-changing routine.
Move it into its own function, and have it set the base type associated
with the color.  This means that the hacks in draco_subspecies() to get
Tiamat's base type and resistances based on color are no longer
necessary, so they're removed.

As a side effect, Tiamat is now always initialized to one of her random
colors instead of always starting out as a purple draconian; if this
were not the case, her base type would be the base draconian until her
first color change, since her genus and species are both set to
MONS_DRACONIAN.