New option tile_align_at_top (windows only)
When set, the window will be aligned at the top of the screen instead of being centered (avoid taskbar overlap, see mantis 2043).

In theory, it would be possible to give the player full control of the window position, but it's more complicated and I don't see any use for it. It can easily be added if anyone ever request it.