Convert grey draconians' unbreathing into a general quality.
On the player side, it's applied specifically to mature grey draconians
and generically to undead races (in both cases as a single-level
non-random mutation), and also applied as a quality to lich and statue
transformations.  On the monster side, it's a monster class flag,
applied specifically to grey draconians and generically to all undead,
plant, and non-living monsters.

As with grey draconians, having this quality implies asphyxiation
resistance and Mephitic Cloud immunity.