New monster: sky beasts. Electric/air versions of ice beasts.
They flicker in and out of visibility roughly every eight turns. This is
clearly not a problem for races or character builds with see
invisibility, but it does provide an interesting effect for those
without.

Currently, they appear about the same levels as ice beasts. This could
be problematic as ice beasts appear quite early, and they don't have the
same vulnerabilities as ice beasts. This could just require their HP
being tweaked down, or something.

We'll see!

They need a tile, too.