Allow variable mutation scheduling within a facet
Facets now have an overall tier (this is the actual tier of the
facet and is only used to pick mutations according to the values set
forth in ct_of_tier[]) as well as the tiers array which represents the
tiers of each mutation which compose the facet. The values in the
tiers array should not exceed the overall tier of the facet.