Fix staves not fully iding under Ashenzari.
The bug was in them getting pluses on a tier 2 id, but not getting the base
type ided on tier 1.  I could make it consistent with jewelry and making only
self-iding staves identify on tier 1, but then, it's a matter of casting any
spell ~10-20 times, so I just made it reveal the type early.  Rods still get
just the type on tier 1 and enchantment on 2.