Mark smoke demons as unbreathing.
They're described as being made out of smoke, and marked as
insubstantial (the only demons marked as insubstantial, in fact) and
poison-resistant.  The first two of these points to their not having any
breathing apparatus (which is important, since demons no longer
automatically resist asphyxiation), and the third makes them immune to
Mephitic Cloud, which unbreathing covers anyway.