Get rid of the old species_skill() function
This function returned the aptitude value of the old system. I Replaced it with
species_apt() or species_apt_factor(). This simplify the code in several places.

It also reduce rounding error in skill_exp_needed() calculation which has the
funny side effect that you can lose a skill level for a char created before the
change. The good news is you'll regain it very soon (it's only a single point)