Properly set an upper limit for fresh undead abominations' AC bonuses.
There was no limit before, unlike hit dice, which could cause problems
if enough fresh corpses were used (unlikely, but possible).

The current upper limit for abominations with bonuses is 20 AC, since
the upper limit for their HD is 30, and the highest AC to be found on
monsters with 30 HD (e.g. orbs of fire) is 20.