A command-line argument --print-charset to query the i/o encoding Crawl wants.
This is to let DGL support both Crawls that obey system locale settings and
ones which do not.

I guess we'll use another approach and this will be reverted, but if not,
having this commit in would reduce the amount of incompatible builds on CDO
(as it's currently impossible to update past trunk versions).