Add cloud with disturbance tiles for invisible monsters in cloud.
This fix Mantis 2573.

I didn't do ink, forest fire, holy flames and gloom because even though they are
supposed to be opaque clouds, it seemed unnecessary.

To make the tiles, I used the unseen monster tile, filed it with 50% gray,
blurred it and used it as a layer mask to add transparency to the clouds.
I saved the shape as UNUSED/other/cloud_disturbance_shape.png
This is my first tiles, so someone more skilled should be able to do better.