Detect portals on level when joining Ashenzari.
There's a question about when portal detection should start -- perhaps a
piety threshold would be better, but in any case there should be no need
to leave/enter the level to detect not yet discovered portals there.

Sewers, Ossuaries and even Baileys are likely to occur in the overflow
temple range, and other altars can appear anywhere (with Trovel, literally).