Remove rune_subst; bump save major version, map cache version; clean up after version bump.
This is rather large, but rune_subst was in the saves, and I decided it was
cleaner to bump the major version than the minor one.

This updates enum.h.  I apologize for the inconvenience.

This might have broken something, but the previous changes to .des files should
have taken care of that.