Remove use of mons_global_level to filter results from ?/M.
Also removes mons_global_level itself, since it is no longer used anywhere, and
other things from mon-pick.cc that were only used by it.

The filtering by mons_global_level seems to only have effect when the game is
not started, that is, when ?/M is used on the character selection screen.  When
the game is running, its results are the same as without the mons_global_level
filter.  So there's no point to keep it around.