Make teleport use is_feat_dangerous to exclude cells.
This simplify the function a lot and allow grey drac to teleport into deep
water and permanent flyers (kenku and boots of levitation) to teleport over
deep water and lava.
On the other hand, as a merfolk in a non-swimming form, you can't teleport
over deep water and quickly turn back, but I think this is ok. And if we want
to bring it back, let's do it in a more generic way (ie compatible with grey
draconians).