Get rid of you.attribute[ATTR_TRANSFORMATION].
Galehar's your_name hack was a step in the right direction, but it was still
not fully clean.  There's no reason to use you.attribute[] -- it has a benefit
of not requiring a minor tag (something that was even impossible in the past!),
but has its flaws like verbosity, being hard to read and not allowing using
proper types the compiler can use to give warnings.