Allow the player more control over what items to sac to Nemelex.
Implement a toggle on the second religion screen (^!) that allows
fine-tuning over which item types to skip and which ones to actually
sacrifice when praying.

So far, this behaviour is completely undocumented.

There's currently no feedback at all if prayer skips over items or
even if everything in a stack turns out not to match the player's
sacrifice categories.
Also, we probably don't want such ineffective prayer to cost a turn.