New skill menu
Features
- Contextual 2 lines help.
- 1 line footer, with current commands
- Can navigate the menu with arrow keys and select skills with enter.
- In tiles, everything can be operated with the mouse.
- Aptitudes are always showned (can be made optional).
- Reskilling is a bit more convenient.
- New features are now much easier to add.

Known bugs and limitation
- Tiles: the whole display is shifted about 2 lines down and 2 columns to the
right. The mouse interacts with items where they are supposed to be, so it's
almost unusable (this might be related to 87b100a1, or not).
- Tiles: If you click to leave a description, the clic isn't eaten, so you
might immediately bring another (or the same) description.
- Console: the highligting doesn't handle properly the change of text colour.
Toggling the practise of a skill with enter turn its name invisible.
- Console: sometimes, the default colour for the help changes.
- The new FormattedTextItem ignores fg_colour (only in tiles). If we fix that
we can merge the new class with TextItem providing multicolour to all derived
classes. Using formatted_string also simplify the render method.
- You can't change column if there isn't a selectable skill right in front of
the current active one. It should move to the nearest selectable one.
- I've added a new header file (skills_menu.h) just to reduce the number of
recompiling while it's beta. It should probably be merged in skills2.h before
the whole branch get merged in trunk.