Prevent _roguey_level from being overwritten by _builder_basic.
This removes levels that are generated by both the builders.  There used to be a
50% chance of any roguey_level to be overwritten by _builder_basic.  The
rationale is that it makes the code path much cleaner; additionally, levels
generated in this manner are usually not very remarkable, often having little in
common with either of the generators employed (_builder_basic tends to overwrite
a good deal of the level).