Convert special rooms to a minivault that uses subvaults.
rooms.des' contents become subvaults, used by the 'special_room' minivault.

The behaviour is now quite different: The room is connected to the rest of the
level just like any minivault, which in most cases means short or no corridor.
So the room now always gets a door.

Placement (DEPTH: 6-30, CHANCE: 20%) should probably be readjusted, but that is
quite easy now.

Special room building code in _roguey_level has been commented out, so it cannot
build special rooms at the moment.

Aside from that, effort has been put into preserving at least the distribution
of different types of special rooms in the dungeon.  Some of the previous
changes have been made to aid this.

I am, as always, not really confident in my lua code, and subvault-related code
in rooms.des in particular.  Somebody with knowledge of the topic should look
into it and fix it.