On big terminals and windows, keep the skill menu at 80x24 and center it.
This is experimental. All the menus in crawl are on the top left. There should
be some consistency, so either this one goes to the corner, or the others are
centered. I'm not sure about console, but int tiles with a big monitor,
having menues in the corner doesn't look very good.