Set the license info to what we'll probably end up with.
I need to add an outside file with a different license, and this is
forbidden by the current old-Bison crap we have, even for extremely
permissive licenses.  We already are in breach by including mt19937 and
stdint.h -- both under BSD3 which is not the "Crawl license".  That's the
reason why we need GPL, BSD or anything which allows "compatible licenses".

Fortunately, the ongoing relicensing effort is planned to complete before
0.8, in the same time frame as this branch.  Thus, I'm already marking it
GPL since the current license is no good.

This means the branch is undistributable until we get this finished.