Revert "Fix 'radiating silence' bug", it relied on undefined behaviour.
Quoting the C++ standard:

"The operands shall be of integral or unscoped enumeration type and integral
promotions are performed.  The type of the result is that of the promoted
left operand.  The behavior is undefined if the right operand is negative,
or greater than or equal to the length in bits of the promoted left operand."

Thus, the "fixed" version ever working relies on overflow being or not trapped.
The result is flags either being wrapped around to their enum minus 32, or
silently lost.

The proper solution is to have the left operand have the needed width.