Double GDR from body armour.
Limited testing suggests that plate is now on average as good protection as a robe against a stone giant, and a lot better against an executioner.

With this, enchantments cap changes and removal of AC/EV bonuses from randarts, there should be plenty to test regarding heavy armour characters. Happy crawling!