Fix [2941], a crash involving kraken heads killing themselves
Fix a crash involving krakens killing their 'head' segment after
creating a tentacle with mindex 0. The check used to determine if a
monster was a connector to be deleted allowed kraken heads, tentacles,
and connectors rather than just connectors, so if a tentacle was created
with mindex 0, the head would get deleted (since the heads number is 0
by default).