Recite resistance tweaking; there are less bad effects early on, and it is still useful later.
Modeled after 6d17f6c1b, but less overpowered. That seemed fun and enjoyable, even if broken.

Early game: with Priests of Zin, there are practically no bad effects, and you can Recite quite freely to lone kobolds, goblins and hobgoblins. This will not work against orc packs with wizards and priests, however.

If you pick Zin on an altar, with low Invocations Recite you get some bad effects from the puny humanoids, and sometimes you succeed.

In late game, there are monsters against which Recite is useful, but on a bad roll they may get angered. And then there are monsters, against which it isn't useful, but which get angered when you Recite with them around.

I'm committing this to testing, to see if the current Recite is salvageable with number tweaking. See #2850 for a more comprehensive Recite patch project, which may well be worth implementing even if this works.